iou: fix not exiting io_uring IO engine
Graceful shutdown relies on recv() returning 0 for the main_sock after
test ends i.e. self->ended = 1. But I'm returning early in that case
thus never exiting the worker loop.

The final cancel when closing a conn is sync i.e. by the time submit
returns the request has been cancelled. But it still generates a
completion. Add the missing handler as to not spam with error msg.

Signed-off-by: David Wei <dw@davidwei.uk>